refactor(admin): extract applyBlocks handler
